[UPD] Added input sending to the server

+++ client/NetworkECSMediator.cpp
@@ -1,5 +1,6 @@
 #include "NetworkECSMediator.hpp"
 #include "Network/Receiver.hpp"
+#include <cstdint>
 #include <exception>
 #include <iostream>
 #include <stdexcept>
@@ -8,22 +9,22 @@ NetworkECSMediator::NetworkECSMediator()
 {
     _mediatorMap = {
         {static_cast<int>(NetworkECSMediatorEvent::SEND_DATA_TCP),
-         [this](const std::string &data, int opcode) {
+         [this](const std::string &data, uint8_t opcode) {
              std::cout << "SENDING TCP" << std::endl;
              _sender->sendTcp(opcode, data);
          }},
         {static_cast<int>(NetworkECSMediatorEvent::SEND_DATA_UDP),
-         [this](const std::string &data, int opcode) {
+         [this](const std::string &data, uint8_t opcode) {
              std::cout << "SENDING UDP" << std::endl;
              _sender->sendUdp(opcode, data);
          }},
-        {static_cast<int>(NetworkECSMediatorEvent::UPDATE_DATA), [this](const std::string &data, int opcode) {
+        {static_cast<int>(NetworkECSMediatorEvent::UPDATE_DATA), [this](const std::string &data, uint8_t opcode) {
              std::cout << "RECEIVED DATA: " << data << std::endl;
              // Call ECS update functions here
          }}};
 }
 
-void NetworkECSMediator::notify(NetworkECSMediatorEvent event, const std::string &data, int opcode)
+void NetworkECSMediator::notify(NetworkECSMediatorEvent event, const std::string &data, uint8_t opcode)
 {
     auto it = _mediatorMap.end();
     try
